refactor(storage): extract WebhookRepository from storage adapters
Addresses the TODO on postgres.adapter.ts L49. Extracts all webhook and delivery methods from PostgresAdapter, SqliteAdapter, and MemoryAdapter into dedicated WebhookRepository classes under apps/api/src/storage/adapters/repositories/.

Each adapter becomes a thin facade that delegates to its repository. Zero behavior changes - existing webhook-storage test suite (25 tests) passes unchanged.

Establishes the pattern for extracting the remaining domains (ACL, anomaly, slowlog, commandlog, latency, memory, hotkeys, etc.) in follow-up PRs.

Made-with: Cursor
