+++ apps/api/src/storage/adapters/postgres.adapter.ts
@@ -45,9 +45,11 @@ import type {
   MetricKind,
 } from '@betterdb/shared';
 import { PostgresDialect, RowMappers } from './base-sql.adapter';
+import { WebhookPostgresRepository } from './repositories/webhook.postgres.repository';
 
-// TODO: Split into domain-specific repositories (acl, webhooks, anomaly, slowlog, etc.)
-// and turn PostgresAdapter into a thin facade that delegates to them.
+// Domain-specific repositories (webhooks extracted). Remaining domains to extract:
+// ACL, anomaly, slowlog, commandlog, latency, memory, hotkeys, settings,
+// agent-tokens, metric-forecasts, client-snapshots, key-patterns, vector-index-snapshots.
 
 export interface PostgresAdapterConfig {
   connectionString: string;
@@ -58,6 +60,7 @@ export class PostgresAdapter implements StoragePort {
   private pool: Pool | null = null;
   private ready: boolean = false;
   private readonly mappers = new RowMappers(PostgresDialect);
+  private webhookRepo!: WebhookPostgresRepository;
 
   constructor(private config: PostgresAdapterConfig) {}
 
@@ -197,6 +200,8 @@ export class PostgresAdapter implements StoragePort {
           });
         });
       }
+
+      this.webhookRepo = new WebhookPostgresRepository(this.pool, this.mappers);
 
       // Test connection (will have correct search_path if schema is set)
       const testClient = await this.pool.connect();
@@ -2324,75 +2329,22 @@ export class PostgresAdapter implements StoragePort {
 
   async createWebhook(webhook: Omit<Webhook, 'id' | 'createdAt' | 'updatedAt'>): Promise<Webhook> {
     if (!this.pool) throw new Error('Database not initialized');
-
-    const result = await this.pool.query(
-      `INSERT INTO webhooks (name, url, secret, enabled, events, headers, retry_policy, delivery_config, alert_config, thresholds, connection_id)
-       VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11)
-       RETURNING *`,
-      [
-        webhook.name,
-        webhook.url,
-        webhook.secret,
-        webhook.enabled,
-        webhook.events,
-        JSON.stringify(webhook.headers || {}),
-        JSON.stringify(webhook.retryPolicy),
-        webhook.deliveryConfig ? JSON.stringify(webhook.deliveryConfig) : null,
-        webhook.alertConfig ? JSON.stringify(webhook.alertConfig) : null,
-        webhook.thresholds ? JSON.stringify(webhook.thresholds) : null,
-        webhook.connectionId || null,
-      ],
-    );
-
-    return this.mappers.mapWebhookRow(result.rows[0]);
+    return this.webhookRepo.createWebhook(webhook);
   }
 
   async getWebhook(id: string): Promise<Webhook | null> {
     if (!this.pool) throw new Error('Database not initialized');
-
-    const result = await this.pool.query('SELECT * FROM webhooks WHERE id = $1', [id]);
-    if (result.rows.length === 0) return null;
-
-    return this.mappers.mapWebhookRow(result.rows[0]);
+    return this.webhookRepo.getWebhook(id);
   }
 
   async getWebhooksByInstance(connectionId?: string): Promise<Webhook[]> {
     if (!this.pool) throw new Error('Database not initialized');
-
-    if (connectionId) {
-      const result = await this.pool.query(
-        'SELECT * FROM webhooks WHERE connection_id = $1 OR connection_id IS NULL ORDER BY created_at DESC',
-        [connectionId],
-      );
-      return result.rows.map((row) => this.mappers.mapWebhookRow(row));
-    }
-
-    // No connectionId provided - only return global webhooks (not scoped to any connection)
-    const result = await this.pool.query(
-      'SELECT * FROM webhooks WHERE connection_id IS NULL ORDER BY created_at DESC',
-    );
-    return result.rows.map((row) => this.mappers.mapWebhookRow(row));
+    return this.webhookRepo.getWebhooksByInstance(connectionId);
   }
 
   async getWebhooksByEvent(event: WebhookEventType, connectionId?: string): Promise<Webhook[]> {
     if (!this.pool) throw new Error('Database not initialized');
-
-    if (connectionId) {
-      // Return webhooks scoped to this connection OR global webhooks (no connectionId)
-      const result = await this.pool.query(
-        'SELECT * FROM webhooks WHERE enabled = true AND $1 = ANY(events) AND (connection_id = $2 OR connection_id IS NULL)',
-        [event, connectionId],
-      );
-      return result.rows.map((row) => this.mappers.mapWebhookRow(row));
-    }
-
-    // No connectionId provided - only return global webhooks (not scoped to any connection)
-    const result = await this.pool.query(
-      'SELECT * FROM webhooks WHERE enabled = true AND $1 = ANY(events) AND connection_id IS NULL',
-      [event],
-    );
-
-    return result.rows.map((row) => this.mappers.mapWebhookRow(row));
+    return this.webhookRepo.getWebhooksByEvent(event, connectionId);
   }
 
   async updateWebhook(
@@ -2400,117 +2352,24 @@ export class PostgresAdapter implements StoragePort {
     updates: Partial<Omit<Webhook, 'id' | 'createdAt' | 'updatedAt'>>,
   ): Promise<Webhook | null> {
     if (!this.pool) throw new Error('Database not initialized');
-
-    const setClauses: string[] = [];
-    const params: any[] = [];
-    let paramIndex = 1;
-
-    if (updates.name !== undefined) {
-      setClauses.push(`name = $${paramIndex++}`);
-      params.push(updates.name);
-    }
-    if (updates.url !== undefined) {
-      setClauses.push(`url = $${paramIndex++}`);
-      params.push(updates.url);
-    }
-    if (updates.secret !== undefined) {
-      setClauses.push(`secret = $${paramIndex++}`);
-      params.push(updates.secret);
-    }
-    if (updates.enabled !== undefined) {
-      setClauses.push(`enabled = $${paramIndex++}`);
-      params.push(updates.enabled);
-    }
-    if (updates.events !== undefined) {
-      setClauses.push(`events = $${paramIndex++}`);
-      params.push(updates.events);
-    }
-    if (updates.headers !== undefined) {
-      setClauses.push(`headers = $${paramIndex++}`);
-      params.push(JSON.stringify(updates.headers));
-    }
-    if (updates.retryPolicy !== undefined) {
-      setClauses.push(`retry_policy = $${paramIndex++}`);
-      params.push(JSON.stringify(updates.retryPolicy));
-    }
-    if (updates.deliveryConfig !== undefined) {
-      setClauses.push(`delivery_config = $${paramIndex++}`);
-      params.push(JSON.stringify(updates.deliveryConfig));
-    }
-    if (updates.alertConfig !== undefined) {
-      setClauses.push(`alert_config = $${paramIndex++}`);
-      params.push(JSON.stringify(updates.alertConfig));
-    }
-    if (updates.thresholds !== undefined) {
-      setClauses.push(`thresholds = $${paramIndex++}`);
-      params.push(JSON.stringify(updates.thresholds));
-    }
-    if (updates.connectionId !== undefined) {
-      setClauses.push(`connection_id = $${paramIndex++}`);
-      params.push(updates.connectionId);
-    }
-
-    if (setClauses.length === 0) {
-      return this.getWebhook(id);
-    }
-
-    setClauses.push(`updated_at = NOW()`);
-    params.push(id);
-
-    const result = await this.pool.query(
-      `UPDATE webhooks SET ${setClauses.join(', ')} WHERE id = $${paramIndex} RETURNING *`,
-      params,
-    );
-
-    if (result.rows.length === 0) return null;
-
-    return this.mappers.mapWebhookRow(result.rows[0]);
+    return this.webhookRepo.updateWebhook(id, updates);
   }
 
   async deleteWebhook(id: string): Promise<boolean> {
     if (!this.pool) throw new Error('Database not initialized');
-
-    const result = await this.pool.query('DELETE FROM webhooks WHERE id = $1', [id]);
-    return (result.rowCount ?? 0) > 0;
+    return this.webhookRepo.deleteWebhook(id);
   }
 
   async createDelivery(
     delivery: Omit<WebhookDelivery, 'id' | 'createdAt'>,
   ): Promise<WebhookDelivery> {
     if (!this.pool) throw new Error('Database not initialized');
-
-    const result = await this.pool.query(
-      `INSERT INTO webhook_deliveries (
-        webhook_id, event_type, payload, status, status_code, response_body,
-        attempts, next_retry_at, completed_at, duration_ms, connection_id
-      )
-      VALUES ($1, $2, $3, $4, $5, $6, $7, $8, $9, $10, $11)
-      RETURNING *`,
-      [
-        delivery.webhookId,
-        delivery.eventType,
-        JSON.stringify(delivery.payload),
-        delivery.status,
-        delivery.statusCode || null,
-        delivery.responseBody || null,
-        delivery.attempts,
-        delivery.nextRetryAt ? new Date(delivery.nextRetryAt) : null,
-        delivery.completedAt ? new Date(delivery.completedAt) : null,
-        delivery.durationMs || null,
-        delivery.connectionId || null,
-      ],
-    );
-
-    return this.mappers.mapDeliveryRow(result.rows[0]);
+    return this.webhookRepo.createDelivery(delivery);
   }
 
   async getDelivery(id: string): Promise<WebhookDelivery | null> {
     if (!this.pool) throw new Error('Database not initialized');
-
-    const result = await this.pool.query('SELECT * FROM webhook_deliveries WHERE id = $1', [id]);
-    if (result.rows.length === 0) return null;
-
-    return this.mappers.mapDeliveryRow(result.rows[0]);
+    return this.webhookRepo.getDelivery(id);
   }
 
   async getDeliveriesByWebhook(
@@ -2519,13 +2378,7 @@ export class PostgresAdapter implements StoragePort {
     offset: number = 0,
   ): Promise<WebhookDelivery[]> {
     if (!this.pool) throw new Error('Database not initialized');
-
-    const result = await this.pool.query(
-      'SELECT * FROM webhook_deliveries WHERE webhook_id = $1 ORDER BY created_at DESC LIMIT $2 OFFSET $3',
-      [webhookId, limit, offset],
-    );
-
-    return result.rows.map((row) => this.mappers.mapDeliveryRow(row));
+    return this.webhookRepo.getDeliveriesByWebhook(webhookId, limit, offset);
   }
 
   async updateDelivery(
@@ -2533,50 +2386,7 @@ export class PostgresAdapter implements StoragePort {
     updates: Partial<Omit<WebhookDelivery, 'id' | 'webhookId' | 'createdAt'>>,
   ): Promise<boolean> {
     if (!this.pool) throw new Error('Database not initialized');
-
-    const setClauses: string[] = [];
-    const params: any[] = [];
-    let paramIndex = 1;
-
-    if (updates.status !== undefined) {
-      setClauses.push(`status = $${paramIndex++}`);
-      params.push(updates.status);
-    }
-    if (updates.statusCode !== undefined) {
-      setClauses.push(`status_code = $${paramIndex++}`);
-      params.push(updates.statusCode);
-    }
-    if (updates.responseBody !== undefined) {
-      setClauses.push(`response_body = $${paramIndex++}`);
-      params.push(updates.responseBody);
-    }
-    if (updates.attempts !== undefined) {
-      setClauses.push(`attempts = $${paramIndex++}`);
-      params.push(updates.attempts);
-    }
-    if (updates.nextRetryAt !== undefined) {
-      setClauses.push(`next_retry_at = $${paramIndex++}`);
-      params.push(updates.nextRetryAt ? new Date(updates.nextRetryAt) : null);
-    }
-    if (updates.completedAt !== undefined) {
-      setClauses.push(`completed_at = $${paramIndex++}`);
-      params.push(updates.completedAt ? new Date(updates.completedAt) : null);
-    }
-    if (updates.durationMs !== undefined) {
-      setClauses.push(`duration_ms = $${paramIndex++}`);
-      params.push(updates.durationMs);
-    }
-
-    if (setClauses.length === 0) return true;
-
-    params.push(id);
-
-    const result = await this.pool.query(
-      `UPDATE webhook_deliveries SET ${setClauses.join(', ')} WHERE id = $${paramIndex}`,
-      params,
-    );
-
-    return (result.rowCount ?? 0) > 0;
+    return this.webhookRepo.updateDelivery(id, updates);
   }
 
   async getRetriableDeliveries(
@@ -2584,51 +2394,12 @@ export class PostgresAdapter implements StoragePort {
     connectionId?: string,
   ): Promise<WebhookDelivery[]> {
     if (!this.pool) throw new Error('Database not initialized');
-
-    if (connectionId) {
-      const result = await this.pool.query(
-        `SELECT * FROM webhook_deliveries
-         WHERE status = 'retrying'
-         AND next_retry_at <= NOW()
-         AND connection_id = $2
-         ORDER BY next_retry_at ASC
-         LIMIT $1
-         FOR UPDATE SKIP LOCKED`,
-        [limit, connectionId],
-      );
-      return result.rows.map((row) => this.mappers.mapDeliveryRow(row));
-    }
-
-    const result = await this.pool.query(
-      `SELECT * FROM webhook_deliveries
-       WHERE status = 'retrying'
-       AND next_retry_at <= NOW()
-       ORDER BY next_retry_at ASC
-       LIMIT $1
-       FOR UPDATE SKIP LOCKED`,
-      [limit],
-    );
-
-    return result.rows.map((row) => this.mappers.mapDeliveryRow(row));
+    return this.webhookRepo.getRetriableDeliveries(limit, connectionId);
   }
 
   async pruneOldDeliveries(cutoffTimestamp: number, connectionId?: string): Promise<number> {
     if (!this.pool) throw new Error('Database not initialized');
-
-    if (connectionId) {
-      const result = await this.pool.query(
-        'DELETE FROM webhook_deliveries WHERE EXTRACT(EPOCH FROM created_at) * 1000 < $1 AND connection_id = $2',
-        [cutoffTimestamp, connectionId],
-      );
-      return result.rowCount ?? 0;
-    }
-
-    const result = await this.pool.query(
-      'DELETE FROM webhook_deliveries WHERE EXTRACT(EPOCH FROM created_at) * 1000 < $1',
-      [cutoffTimestamp],
-    );
-
-    return result.rowCount ?? 0;
+    return this.webhookRepo.pruneOldDeliveries(cutoffTimestamp, connectionId);
   }
 
   // Slow Log Methods
